# packages/likec4/src/agent/openapi.yaml
openapi: 3.1.0
info:
  title: LikeC4 AI Agent API
  description: |
    AI Agent for LikeC4 architecture modeling with LiteLLM proxy integration
    and an extensible skills and plugin system.

    The agent backend is started automatically when the `LIKEC4_AGENT_URL`
    environment variable is set (pointing to a LiteLLM or OpenAI-compatible proxy).

    Environment variables:
      - LIKEC4_AGENT_URL   — LiteLLM proxy base URL (required to enable the agent)
      - LIKEC4_AGENT_KEY   — API key (optional, defaults to "no-key")
      - LIKEC4_AGENT_MODEL — Model name (optional, defaults to "gpt-4o")
      - LIKEC4_AGENT_PORT  — Agent server port (optional, defaults to 33336)
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: /agent
    description: LikeC4 Agent endpoint (proxied by the dev server at /agent/*)
paths:
...
  /sessions/{sessionId}:
    get:
      operationId: getSession
      summary: Get conversation session history
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session data with message history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteSession
      summary: Clear a conversation session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Session cleared successfully
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Session:
      type: object
      required: [id, messages]
      properties:
        id:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Message:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
...
  responses:
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: string
...